Move crack-related methods and tests into cracker files. Test cracker exists
